+++ history.md
@@ -1,8 +1,16 @@
 
+# V03.
+
+- DefaultResultBuilder is enriched: method __ok__ without args is added
+- DefaultResultBuilder is enriched: methods __compute__ is added
+- Util. class Seeds is added
+
 # V0.2
+
 - static method fail(Seed)<Result<T>> is added into DefaultResultBuilder
 
 # V0.1
+
 - DefaultSeed is added
 - DefaultResult is added
 - DefaultI18nEngine id added